refactor: extract endpoint models
